requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: for future enhancements
# pydantic>=2.0.0  # If switching from dataclasses
